chore: forge fmt tests (#1212)
**Motivation:**  

Our test suite is growing rapidly, and its formatting has become
increasingly inconsistent.

**Modifications:**  

- Enabled the `test` profile for `forge fmt` (`FOUNDRY_PROFILE=test
forge fmt`).
- Added `make fmt` to format both contracts and tests.  

**Result:**  

Tests and scripts are now consistently formatted, reducing LOC by 3-4k.

+++ script/releases/Env.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.12;
 
 import "forge-std/Vm.sol";
-import "zeus-templates/utils/ZEnvHelpers.sol";  
+import "zeus-templates/utils/ZEnvHelpers.sol";
 
 import {TimelockController} from "@openzeppelin/contracts/governance/TimelockController.sol";
 import "@openzeppelin/contracts/proxy/beacon/UpgradeableBeacon.sol";
@@ -35,14 +35,21 @@ import "src/contracts/token/Eigen.sol";
 import "src/contracts/token/BackingEigen.sol";
 
 library Env {
-
     using ZEnvHelpers for *;
 
     /// Dummy types and variables to facilitate syntax, e.g: `Env.proxy.delegationManager()`
-    enum DeployedProxy { A }
-    enum DeployedBeacon { A }
-    enum DeployedImpl { A }
-    enum DeployedInstance { A }
+    enum DeployedProxy {
+        A
+    }
+    enum DeployedBeacon {
+        A
+    }
+    enum DeployedImpl {
+        A
+    }
+    enum DeployedInstance {
+        A
+    }
 
     DeployedProxy internal constant proxy = DeployedProxy.A;
     DeployedBeacon internal constant beacon = DeployedBeacon.A;
@@ -52,7 +59,6 @@ library Env {
     /**
      * env
      */
-
     function executorMultisig() internal view returns (address) {
         return _envAddress("executorMultisig");
     }
@@ -82,7 +88,7 @@ library Env {
     }
 
     function multiSendCallOnly() internal view returns (address) {
-        return _envAddress("MultiSendCallOnly");   
+        return _envAddress("MultiSendCallOnly");
     }
 
     function EIGENPOD_GENESIS_TIME() internal view returns (uint64) {
@@ -129,195 +135,265 @@ library Env {
         return _envU32("REWARDS_COORDINATOR_GENESIS_REWARDS_TIMESTAMP");
     }
 
-    function REWARDS_PAUSE_STATUS() internal view returns (uint) {
+    function REWARDS_PAUSE_STATUS() internal view returns (uint256) {
         return _envU256("REWARDS_COORDINATOR_PAUSE_STATUS");
     }
 
     /**
      * core/
      */
-
-    function allocationManager(DeployedProxy) internal view returns (AllocationManager) {
+    function allocationManager(
+        DeployedProxy
+    ) internal view returns (AllocationManager) {
         return AllocationManager(_deployedProxy(type(AllocationManager).name));
     }
 
-    function allocationManager(DeployedImpl) internal view returns (AllocationManager) {
+    function allocationManager(
+        DeployedImpl
+    ) internal view returns (AllocationManager) {
         return AllocationManager(_deployedImpl(type(AllocationManager).name));
     }
 
-    function avsDirectory(DeployedProxy) internal view returns (AVSDirectory) {
+    function avsDirectory(
+        DeployedProxy
+    ) internal view returns (AVSDirectory) {
         return AVSDirectory(_deployedProxy(type(AVSDirectory).name));
     }
 
-    function avsDirectory(DeployedImpl) internal view returns (AVSDirectory) {
+    function avsDirectory(
+        DeployedImpl
+    ) internal view returns (AVSDirectory) {
         return AVSDirectory(_deployedImpl(type(AVSDirectory).name));
     }
 
-    function delegationManager(DeployedProxy) internal view returns (DelegationManager) {
+    function delegationManager(
+        DeployedProxy
+    ) internal view returns (DelegationManager) {
         return DelegationManager(_deployedProxy(type(DelegationManager).name));
     }
 
-    function delegationManager(DeployedImpl) internal view returns (DelegationManager) {
-        return DelegationManager(_deployedImpl(type(DelegationManager).name)); 
+    function delegationManager(
+        DeployedImpl
+    ) internal view returns (DelegationManager) {
+        return DelegationManager(_deployedImpl(type(DelegationManager).name));
     }
 
-    function rewardsCoordinator(DeployedProxy) internal view returns (RewardsCoordinator) {
+    function rewardsCoordinator(
+        DeployedProxy
+    ) internal view returns (RewardsCoordinator) {
         return RewardsCoordinator(_deployedProxy(type(RewardsCoordinator).name));
     }
 
-    function rewardsCoordinator(DeployedImpl) internal view returns (RewardsCoordinator) {
-        return RewardsCoordinator(_deployedImpl(type(RewardsCoordinator).name)); 
+    function rewardsCoordinator(
+        DeployedImpl
+    ) internal view returns (RewardsCoordinator) {
+        return RewardsCoordinator(_deployedImpl(type(RewardsCoordinator).name));
     }
 
-    function strategyManager(DeployedProxy) internal view returns (StrategyManager) {
+    function strategyManager(
+        DeployedProxy
+    ) internal view returns (StrategyManager) {
         return StrategyManager(_deployedProxy(type(StrategyManager).name));
     }
 
-    function strategyManager(DeployedImpl) internal view returns (StrategyManager) {
-        return StrategyManager(_deployedImpl(type(StrategyManager).name)); 
+    function strategyManager(
+        DeployedImpl
+    ) internal view returns (StrategyManager) {
+        return StrategyManager(_deployedImpl(type(StrategyManager).name));
     }
 
     /**
      * permissions/
      */
-
-    function pauserRegistry(DeployedImpl) internal view returns (PauserRegistry) {
+    function pauserRegistry(
+        DeployedImpl
+    ) internal view returns (PauserRegistry) {
         return PauserRegistry(_deployedImpl(type(PauserRegistry).name));
     }
 
-    function permissionController(DeployedProxy) internal view returns (PermissionController) {
+    function permissionController(
+        DeployedProxy
+    ) internal view returns (PermissionController) {
         return PermissionController(_deployedProxy(type(PermissionController).name));
     }
 
-    function permissionController(DeployedImpl) internal view returns (PermissionController) {
-        return PermissionController(_deployedImpl(type(PermissionController).name)); 
+    function permissionController(
+        DeployedImpl
+    ) internal view returns (PermissionController) {
+        return PermissionController(_deployedImpl(type(PermissionController).name));
     }
 
     /**
      * pods/
      */
-
-    function eigenPod(DeployedBeacon) internal view returns (UpgradeableBeacon) {
+    function eigenPod(
+        DeployedBeacon
+    ) internal view returns (UpgradeableBeacon) {
         return UpgradeableBeacon(_deployedBeacon(type(EigenPod).name));
     }
 
-    function eigenPod(DeployedImpl) internal view returns (EigenPod) {
-        return EigenPod(payable(_deployedImpl(type(EigenPod).name))); 
+    function eigenPod(
+        DeployedImpl
+    ) internal view returns (EigenPod) {
+        return EigenPod(payable(_deployedImpl(type(EigenPod).name)));
     }
 
-    function eigenPodManager(DeployedProxy) internal view returns (EigenPodManager) {
+    function eigenPodManager(
+        DeployedProxy
+    ) internal view returns (EigenPodManager) {
         return EigenPodManager(_deployedProxy(type(EigenPodManager).name));
     }
 
-    function eigenPodManager(DeployedImpl) internal view returns (EigenPodManager) {
-        return EigenPodManager(_deployedImpl(type(EigenPodManager).name)); 
+    function eigenPodManager(
+        DeployedImpl
+    ) internal view returns (EigenPodManager) {
+        return EigenPodManager(_deployedImpl(type(EigenPodManager).name));
     }
 
     /**
      * strategies/
      */
-
-    function eigenStrategy(DeployedProxy) internal view returns (EigenStrategy) {
+    function eigenStrategy(
+        DeployedProxy
+    ) internal view returns (EigenStrategy) {
         return EigenStrategy(_deployedProxy(type(EigenStrategy).name));
     }
 
-    function eigenStrategy(DeployedImpl) internal view returns (EigenStrategy) {
-        return EigenStrategy(_deployedImpl(type(EigenStrategy).name)); 
+    function eigenStrategy(
+        DeployedImpl
+    ) internal view returns (EigenStrategy) {
+        return EigenStrategy(_deployedImpl(type(EigenStrategy).name));
     }
 
     // Beacon proxy
-    function strategyBase(DeployedBeacon) internal view returns (UpgradeableBeacon) {
+    function strategyBase(
+        DeployedBeacon
+    ) internal view returns (UpgradeableBeacon) {
         return UpgradeableBeacon(_deployedBeacon(type(StrategyBase).name));
     }
 
     // Beaconed impl
-    function strategyBase(DeployedImpl) internal view returns (StrategyBase) {
-        return StrategyBase(_deployedImpl(type(StrategyBase).name)); 
+    function strategyBase(
+        DeployedImpl
+    ) internal view returns (StrategyBase) {
+        return StrategyBase(_deployedImpl(type(StrategyBase).name));
     }
 
     // Returns the number of proxy instances
-    function strategyBaseTVLLimits_Count(DeployedInstance) internal view returns (uint) {
+    function strategyBaseTVLLimits_Count(
+        DeployedInstance
+    ) internal view returns (uint256) {
         return _deployedInstanceCount(type(StrategyBaseTVLLimits).name);
     }
 
     // Returns the proxy instance at index `i`
-    function strategyBaseTVLLimits(DeployedInstance, uint i) internal view returns (StrategyBaseTVLLimits) {
+    function strategyBaseTVLLimits(DeployedInstance, uint256 i) internal view returns (StrategyBaseTVLLimits) {
         return StrategyBaseTVLLimits(_deployedInstance(type(StrategyBaseTVLLimits).name, i));
     }
 
-    function strategyBaseTVLLimits(DeployedImpl) internal view returns (StrategyBaseTVLLimits) {
-        return StrategyBaseTVLLimits(_deployedImpl(type(StrategyBaseTVLLimits).name)); 
+    function strategyBaseTVLLimits(
+        DeployedImpl
+    ) internal view returns (StrategyBaseTVLLimits) {
+        return StrategyBaseTVLLimits(_deployedImpl(type(StrategyBaseTVLLimits).name));
     }
 
-    function strategyFactory(DeployedProxy) internal view returns (StrategyFactory) {
+    function strategyFactory(
+        DeployedProxy
+    ) internal view returns (StrategyFactory) {
         return StrategyFactory(_deployedProxy(type(StrategyFactory).name));
     }
 
-    function strategyFactory(DeployedImpl) internal view returns (StrategyFactory) {
-        return StrategyFactory(_deployedImpl(type(StrategyFactory).name)); 
+    function strategyFactory(
+        DeployedImpl
+    ) internal view returns (StrategyFactory) {
+        return StrategyFactory(_deployedImpl(type(StrategyFactory).name));
     }
 
     /**
      * token/
      */
-    
-    function eigen(DeployedProxy) internal view returns (IEigen) {
+    function eigen(
+        DeployedProxy
+    ) internal view returns (IEigen) {
         return IEigen(_deployedProxy(type(Eigen).name));
     }
 
-    function eigen(DeployedImpl) internal view returns (IEigen) {
-        return IEigen(_deployedImpl(type(Eigen).name)); 
+    function eigen(
+        DeployedImpl
+    ) internal view returns (IEigen) {
+        return IEigen(_deployedImpl(type(Eigen).name));
     }
 
-    function beigen(DeployedProxy) internal view returns (IBackingEigen) {
+    function beigen(
+        DeployedProxy
+    ) internal view returns (IBackingEigen) {
         return IBackingEigen(_deployedProxy(type(BackingEigen).name));
     }
 
-    function beigen(DeployedImpl) internal view returns (IBackingEigen) {
-        return IBackingEigen(_deployedImpl(type(BackingEigen).name)); 
+    function beigen(
+        DeployedImpl
+    ) internal view returns (IBackingEigen) {
+        return IBackingEigen(_deployedImpl(type(BackingEigen).name));
     }
 
     /**
      * Helpers
      */
-
-    function _deployedInstance(string memory name, uint idx) private view returns (address) {
+    function _deployedInstance(string memory name, uint256 idx) private view returns (address) {
         return ZEnvHelpers.state().deployedInstance(name, idx);
     }
 
-    function _deployedInstanceCount(string memory name) private view returns (uint) {
+    function _deployedInstanceCount(
+        string memory name
+    ) private view returns (uint256) {
         return ZEnvHelpers.state().deployedInstanceCount(name);
     }
 
-    function _deployedProxy(string memory name) private view returns (address) {
+    function _deployedProxy(
+        string memory name
+    ) private view returns (address) {
         return ZEnvHelpers.state().deployedProxy(name);
     }
 
-    function _deployedBeacon(string memory name) private view returns (address) {
+    function _deployedBeacon(
+        string memory name
+    ) private view returns (address) {
         return ZEnvHelpers.state().deployedBeacon(name);
     }
 
-    function _deployedImpl(string memory name) private view returns (address) {
+    function _deployedImpl(
+        string memory name
+    ) private view returns (address) {
         return ZEnvHelpers.state().deployedImpl(name);
     }
 
-    function _envAddress(string memory key) private view returns (address) {
+    function _envAddress(
+        string memory key
+    ) private view returns (address) {
         return ZEnvHelpers.state().envAddress(key);
     }
 
-    function _envU256(string memory key) private view returns (uint) {
+    function _envU256(
+        string memory key
+    ) private view returns (uint256) {
         return ZEnvHelpers.state().envU256(key);
     }
 
-    function _envU64(string memory key) private view returns (uint64) {
+    function _envU64(
+        string memory key
+    ) private view returns (uint64) {
         return ZEnvHelpers.state().envU64(key);
     }
 
-    function _envU32(string memory key) private view returns (uint32) {
+    function _envU32(
+        string memory key
+    ) private view returns (uint32) {
         return ZEnvHelpers.state().envU32(key);
     }
 
-    function _envU16(string memory key) private view returns (uint16) {
+    function _envU16(
+        string memory key
+    ) private view returns (uint16) {
         return ZEnvHelpers.state().envU16(key);
     }
 }
