+++ src/test/integration/tests/Delegate_Deposit_Queue_Complete.t.sol
@@ -5,15 +5,14 @@ import "src/test/integration/IntegrationChecks.t.sol";
 import "src/test/integration/users/User.t.sol";
 
 contract Integration_Delegate_Deposit_Queue_Complete is IntegrationCheckUtils {
-
     function testFuzz_delegate_deposit_queue_completeAsShares(uint24 _random) public rand(_random) {
         // Create a staker and an operator with a nonzero balance and corresponding strategies
         (User staker, IStrategy[] memory strategies, uint[] memory tokenBalances) = _newRandomStaker();
-        (User operator, ,) = _newRandomOperator();
+        (User operator,,) = _newRandomOperator();
 
         // 1. Delegate to operator
         staker.delegateTo(operator);
-        check_Delegation_State(staker, operator, strategies, new uint256[](strategies.length)); // Initial shares are zero
+        check_Delegation_State(staker, operator, strategies, new uint[](strategies.length)); // Initial shares are zero
 
         // 2. Deposit into strategy
         staker.depositIntoEigenlayer(strategies, tokenBalances);
@@ -40,11 +39,11 @@ contract Integration_Delegate_Deposit_Queue_Complete is IntegrationCheckUtils {
     function testFuzz_delegate_deposit_queue_completeAsTokens(uint24 _random) public rand(_random) {
         // Create a staker and an operator with a nonzero balance and corresponding strategies
         (User staker, IStrategy[] memory strategies, uint[] memory tokenBalances) = _newRandomStaker();
-        (User operator, ,) = _newRandomOperator();
+        (User operator,,) = _newRandomOperator();
 
         // 1. Delegate to operator
         staker.delegateTo(operator);
-        check_Delegation_State(staker, operator, strategies, new uint256[](strategies.length)); // Initial shares are zero
+        check_Delegation_State(staker, operator, strategies, new uint[](strategies.length)); // Initial shares are zero
 
         // 2. Deposit into strategy
         staker.depositIntoEigenlayer(strategies, tokenBalances);
