+++ src/test/integration/tests/Deposit_Delegate_Allocate_Slash_Queue_Redeposit.t.sol
@@ -30,7 +30,7 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
 
         uint[] memory tokensToDeposit = new uint[](initTokenBalances.length);
         numTokensRemaining = new uint[](initTokenBalances.length);
-        uint256 eth_to_deal;
+        uint eth_to_deal;
         for (uint i = 0; i < initTokenBalances.length; i++) {
             if (strategies[i] == BEACONCHAIN_ETH_STRAT) {
                 tokensToDeposit[i] = initTokenBalances[i];
@@ -183,7 +183,9 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             uint[] memory expectedTokens = _calculateExpectedTokens(withdrawals[i].strategies, expectedShares);
             tokens = staker.completeWithdrawalAsTokens(withdrawals[i]);
-            check_Withdrawal_AsTokens_State(staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens);
+            check_Withdrawal_AsTokens_State(
+                staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens
+            );
         }
     }
 
@@ -205,7 +207,9 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             uint[] memory expectedTokens = _calculateExpectedTokens(withdrawals[i].strategies, expectedShares);
             tokens = staker.completeWithdrawalAsTokens(withdrawals[i]);
-            check_Withdrawal_AsTokens_State(staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens);
+            check_Withdrawal_AsTokens_State(
+                staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens
+            );
         }
     }
 
@@ -232,7 +236,9 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             uint[] memory expectedTokens = _calculateExpectedTokens(withdrawals[i].strategies, expectedShares);
             tokens = staker.completeWithdrawalAsTokens(withdrawals[i]);
-            check_Withdrawal_AsTokens_State(staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens);
+            check_Withdrawal_AsTokens_State(
+                staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens
+            );
         }
     }
 
@@ -258,7 +264,9 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             uint[] memory expectedTokens = _calculateExpectedTokens(withdrawals[i].strategies, expectedShares);
             tokens = staker.completeWithdrawalAsTokens(withdrawals[i]);
-            check_Withdrawal_AsTokens_State(staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens);
+            check_Withdrawal_AsTokens_State(
+                staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens
+            );
         }
     }
 
@@ -291,7 +299,9 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             uint[] memory expectedTokens = _calculateExpectedTokens(withdrawals[i].strategies, expectedShares);
             tokens = zeroSharesStaker.completeWithdrawalAsTokens(withdrawals[i]);
-            check_Withdrawal_AsTokens_State(zeroSharesStaker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens);
+            check_Withdrawal_AsTokens_State(
+                zeroSharesStaker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares, tokens, expectedTokens
+            );
         }
     }
 
@@ -307,10 +317,7 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
         check_DecrAlloc_State_Slashable(operator, deallocateParams);
     }
 
-    function testFuzz_fullSlash_undelegate_redeposit_complete(
-        uint24 _random
-    ) public rand(_random) {
-
+    function testFuzz_fullSlash_undelegate_redeposit_complete(uint24 _random) public rand(_random) {
         initDepositShares = _getStakerDepositShares(staker, strategies);
 
         // 4. Fully slash operator
@@ -332,23 +339,19 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
         // 7. Complete withdrawal. Staker should receive 0 shares/tokens after a full slash
         _rollBlocksForCompleteWithdrawals(withdrawals);
 
-        for (uint256 i = 0; i < withdrawals.length; ++i) {
+        for (uint i = 0; i < withdrawals.length; ++i) {
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             staker.completeWithdrawalAsShares(withdrawals[i]);
             check_Withdrawal_AsShares_Undelegated_State(staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares);
         }
-
 
         // Final state checks
         assert_HasExpectedShares(staker, strategies, shares, "staker should have expected shares after redeposit");
         assert_NoWithdrawalsPending(withdrawalRoots, "all withdrawals should be removed from pending");
     }
 
-    function testFuzz_fullSlash_redelegate_redeposit_complete(
-        uint24 _random
-    ) public rand(_random) {
-
-        (User operator2, ,) = _newRandomOperator();
+    function testFuzz_fullSlash_redelegate_redeposit_complete(uint24 _random) public rand(_random) {
+        (User operator2,,) = _newRandomOperator();
         initDepositShares = _getStakerDepositShares(staker, strategies);
 
         // 4. Fully slash operator
@@ -370,12 +373,11 @@ contract Integration_Deposit_Delegate_Allocate_Slash_Queue_Redeposit is Integrat
         // 7. Complete withdrawal. Staker should receive 0 shares/tokens after a full slash
         _rollBlocksForCompleteWithdrawals(withdrawals);
 
-        for (uint256 i = 0; i < withdrawals.length; ++i) {
+        for (uint i = 0; i < withdrawals.length; ++i) {
             uint[] memory expectedShares = _calculateExpectedShares(withdrawals[i]);
             staker.completeWithdrawalAsShares(withdrawals[i]);
             check_Withdrawal_AsShares_Undelegated_State(staker, operator, withdrawals[i], withdrawals[i].strategies, expectedShares);
         }
-
 
         // Final state checks
         assert_HasExpectedShares(staker, strategies, shares, "staker should have expected shares after redeposit");
